genericdatasethandler/src/main/java/com/example/springbatchsupporter/listener/JobCompletionNotificationListener.java
package com.example.springbatchsupporter.listener;

import org.springframework.batch.core.JobExecution;
import org.springframework.batch.core.JobExecutionListener;
import org.springframework.stereotype.Component;

@Component
public class JobCompletionNotificationListener implements JobExecutionListener {

    @Override
    public void beforeJob(JobExecution jobExecution) {
        System.out.println("Job is about to start: " + jobExecution.getJobInstance().getJobName());
    }

    @Override
    public void afterJob(JobExecution jobExecution) {
        System.out.println("Job has completed: " + jobExecution.getJobInstance().getJobName());
    }
}